Adds the draft of the XDP scheduler testing tool
This commit contains the XDP scheduling framework. It consists of a
testing program called xdq-tester used to test schedulers using the XDP
and DEQUEUE hooks. It uses trace files written in Lua that the
xdq-tester program uses to check the XDP schedulers for correctness.

The FIFO, SPRIO, and WFQ are fully functional in this commit. The SPRIO
and WFQ have an API to set the weights from the Lua scripts.

This commit's FQ-CoDel contains sparse flow handling and testing. The
CoDel part has a bug in how it adds the metadata section in this version.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>